@@ -1,8 +1,9 @@
 package rootmulti
 
 import (
-	"github.com/cosmos/cosmos-sdk/store/dbadapter"
-	"github.com/cosmos/cosmos-sdk/store/types"
+	"cosmossdk.io/store/dbadapter"
+	pruningtypes "cosmossdk.io/store/pruning/types"
+	"cosmossdk.io/store/types"
 )
 
 var commithash = []byte("FAKE_HASH")
@@ -30,8 +31,15 @@ func (cdsa commitDBStoreAdapter) LastCommitID() types.CommitID {
 	}
 }
 
-func (cdsa commitDBStoreAdapter) SetPruning(_ types.PruningOptions) {}
+func (cdsa commitDBStoreAdapter) SetPruning(_ pruningtypes.PruningOptions) {}
 
 // GetPruning is a no-op as pruning options cannot be directly set on this store.
 // They must be set on the root commit multi-store.
-func (cdsa commitDBStoreAdapter) GetPruning() types.PruningOptions { return types.PruningOptions{} }
+func (cdsa commitDBStoreAdapter) GetPruning() pruningtypes.PruningOptions {
+	return pruningtypes.NewPruningOptions(pruningtypes.PruningNothing)
+}
+
+// WorkingHash returns the working hash for the adapter
+func (cdsa commitDBStoreAdapter) WorkingHash() []byte {
+	return commithash
+}
